requirements.txt
google-cloud-bigquery>=3.0.0
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=0.20.0